feat : 상수 파일에서 중복되는 badge 상수들을 추가적으로 분리
